Bazı değişkenler int->byte dönüşümü yapılarak alan tasarrufu yapıldı.

+++ ASELSAN_4822.ino
@@ -1,7 +1,8 @@
 #include <Wire.h>
 #include <EEPROM.h>
 #include <avr/pgmspace.h>
-#include "./libraries/fontsandicons.h" 
+#include "./libraries/fontsandicons.h"
+//#include "./libraries/PinChangeInt.h"
 //TODO: join 4822 and 4826 in one code
 //TODO: add PC routines
 //TODO: fix keypad entry speed - Fixed V.1.0b
@@ -68,11 +69,11 @@ byte set_deviceselect = DEVICE_SELECT;
 #define red_led    32
 #define backlight  16
 
-int Led_Status= 240;
+byte Led_Status= 240;
 
-int KeypadIntPin = 4;  //Interrupt Input PIN for MCU
-int KeyVal = 0;     // variable to store the read value
-int old_KeyVal= 0;
+byte KeypadIntPin = 4;  //Interrupt Input PIN for MCU
+byte KeyVal = 0;     // variable to store the read value
+byte old_KeyVal= 0;
 
 #define FWD_POWER_PIN A6
 #define REF_POWER_PIN A7
@@ -101,7 +102,7 @@ int frqSHIFT = 600;
 #define noSHIFT     0
 #define  SIMPLEX    0 //Just incase that we can use this term for noSHIFT
 #define plusSHIFT   1
-int shiftMODE = noSHIFT; // we start with noSHIFT (SIMPLEX)
+byte shiftMODE = noSHIFT; // we start with noSHIFT (SIMPLEX)
 int old_frqSHIFT;       //to store old shift value before entering submenu
 
 
@@ -111,15 +112,15 @@ int old_frqSHIFT;       //to store old shift value before entering submenu
 //Transceiver modes
 #define RX 0
 #define TX 1
-int TRX_MODE = RX; //default transceiver mode is receiving
-int LST_MODE = TX; //this will hold the last receive transmit state. Start with TX because we want to write to PLL on startup 
+byte TRX_MODE = RX; //default transceiver mode is receiving
+byte LST_MODE = TX; //this will hold the last receive transmit state. Start with TX because we want to write to PLL on startup 
 
 
 //RF power control definitions
 #define RF_POWER_PIN A0
 #define HIGH_POWER 0
 #define LOW_POWER  1
-int RF_POWER_STATE = HIGH_POWER; //Initial Power Level is Hight Power
+byte RF_POWER_STATE = HIGH_POWER; //Initial Power Level is Hight Power
 
 
 
@@ -127,7 +128,7 @@ int RF_POWER_STATE = HIGH_POWER; //Initial Power Level is Hight Power
 #define ALERT_PIN 13
 #define ALERT_OFF 0
 #define ALERT_ON  100
-int ALERT_MODE = ALERT_ON;
+byte ALERT_MODE = ALERT_ON;
 
 //Tone Types
 #define NO_tone  0
@@ -140,10 +141,10 @@ int ALERT_MODE = ALERT_ON;
 #define TONE_PIN  3  //D3 is our tone generation PIN (PWM)
 #define CTCSS_OFF 0
 #define CTCSS_ON  1
-int TONE_CTRL = CTCSS_ON; //we start without CTCSS Tone Control
-int ctcss_tone_pos = 8;
+byte TONE_CTRL = CTCSS_ON; //we start without CTCSS Tone Control
+byte ctcss_tone_pos = 8;
 float ctcss_tone_list[50] = {67,69.3,71.9,74.4,77,79.7,82.5,85.4,88.5,91.5,94.8,97.4,100,103.5,107.2,110.9,114.8,118.8,123,127.3,131.8,136.5,141.3,146.3,151.4,156.7,159.8,162.2,165.5,167.9,171.3,173.8,177.3,179.9,183.5,186.2,189.9,192.8,196.6,199.5,203.5,206.5,210.7,218.1,225.7,229.1,233.6,241.8,250.3,254.1};
-int old_ctcss_tone_pos; //to store old tone selection before getting into submenu
+byte old_ctcss_tone_pos; //to store old tone selection before getting into submenu
 
 
 
@@ -153,7 +154,7 @@ int old_ctcss_tone_pos; //to store old tone selection before getting into submen
 #define SQL_ACTIVE 2 //CHANNEL ACTIVE (SQUELCH) PIN
 #define MUTE_PIN_1 6 //PIN for Audio Muting
 //#define MUTE_PIN_2 5
-int CHANNEL_BUSY = 1;
+byte CHANNEL_BUSY = 1;
 
 //MC145158 Programming
 #define pll_clk_pin  9
@@ -170,7 +171,7 @@ char pressedKEY = ' ';
 
 #define scrNORMAL 0
 #define scrMENU   1
-int scrMODE = scrNORMAL;
+byte scrMODE = scrNORMAL;
 
 #define menuNONE   0
 #define menuSQL    1
@@ -178,7 +179,7 @@ int scrMODE = scrNORMAL;
 #define menuSCAN   3
 #define menuRPT    4
 #define menuMENU   5
-int subMENU = menuNONE;
+byte subMENU = menuNONE;
 
 boolean hasASEL = false; //ASELSAN sign
 boolean hasLOCK = false; //KEY LOCK sign
@@ -204,7 +205,7 @@ unsigned char chr2wr[3];
 const char* keymap[4] = {  "123DSX",  "456RB", "789OC", "*0#UM"  };
 const char  numbers[] = "0123456789ABCDEF";
 
-int numChar = 0;
+byte numChar = 0;
 char FRQ[9];// = "145.675 ";
 char FRQ_old[9];// = FRQ;
 long calc_frequency;
@@ -299,7 +300,7 @@ void writeToLcd(const char text[]) {
     //Serial.println(strlen(text),DEC);
     if (idx > 7) break;   
     char *c = strchr(index, (int)toupper(text[idx]));
-    int pos;
+    byte pos;
     if (c == NULL) { 
       pos = 0;      // Char not found, use underscore space instead
     } else {
